[BE]'Merge': "#3.1.1brench Merge, Test 완료"

+++ server_9dokme/src/main/java/com/example/server_9dokme/member/repository/MemberRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Member findBySocialId(String email);
-    Member findByMemberId(int id);
+    Member findByMemberId(Long id);
     Member findByCustomerKey(String customerKey);
     List<Member> findAll();
     Page<Member> findAll(Pageable pageable);
